Shift faculty from last step to step 1
