+++ pages/api/step1.ts
@@ -9,6 +9,7 @@ export const updateApplicationData = (
   contactNo: number,
   currentPosition: string,
   targetProgram: string,
+  faculty: string,
   targetDate: string,
   targetCountry: string,
   enrollProofDoc: string,
@@ -21,6 +22,7 @@ export const updateApplicationData = (
     contact: contactNo,
     current_position: currentPosition,
     target_program: targetProgram,
+    faculty: faculty,
     target_date: targetDate,
     target_country: targetCountry,
     enrollment_proof_doc: enrollProofDoc,
