+++ pages/api/step10.ts
@@ -1,12 +1,8 @@
 import path from 'path'
 import { firestore } from '../../firebase'
 
-export const updateApplicationData = (
-  userId: string,
-  faculty: string,
-  formStatus: number,
-) => {
+export const updateApplicationData = (userId: string, formStatus: number) => {
   return firestore
     .doc(path.join('applications_data', userId))
-    .update({ faculty: faculty, form_status: formStatus })
+    .update({ form_status: formStatus })
 }
